Update 11.7.2013 1057
With the introduction of looping around the azimuthal boundaries within the spreading_hybrid_gaussian and thus accumulating all possible ray contributions on the first instance of creating an eigenray, it no longer became necessary to search for more eigenrays for the specific target at that instant in time and azimuthal angle, and therefore could break out of the search loop and go to the next target.

This assumption is based that only one eigenray can be created for any given combination of azimuthal angle, time, and target. It present it does as required.