Update 12.10.2013 1142
When a target is directly above the source, the assumed corrections for left-handedness fall apart and need further corrections to be taken into account. With the new checks within the spreading_hyrbid logic and is_closest_ray, the wavefront now properly produces a single eigenray when apporiate and at the appropriate time with reasonable values. These new logic introductions show no sign of interference with the az_branch logic or the absence of both logics, aka during beam forming.