Update 12.16.2013 1744
Added new matlab routine for the eigenray_tl_grid/az tests
Updated reflection and refraction tests to produce correct tests
Changed dialog for de_branch on wave_queue.h
Added small debugs to reflection_model.cc
Added clear() to constructor of wvector

NOTE: Need to add default values to constructors. If not, it can cause errors like overflow or underflow, because the memory is allocated for the constructors and the values in the memory are always bogus (ie 2.20958892e-218).