Update 1.8.2014 0926
Installed a new version of Doxygen, 1.8.6, from source and attempted to resolve the issue where enumerations were not showing up in the generated doc files. Upon installing and running Doxygen, we found that the issue still presisted. After working within the Doxy config file for some time, we found that turning on the following parameters resolved the issue: EXTRACT_ALL and EXTRACT_STATIC. After setting these parameters to YES and running Doxygen again, all issues with the docs appeared to be resolved.

Change various files' comments to work properly with Doxygen.