Update 1.8.2014 1138
Updated CMake file to now not show the directories for creating the etopo and woa data files for tests/studies. This was extra information in the CMake gui that wasn't necessary to provide to the user as the files are located within the source data directories.
Moved find_edges call back to public. Was mistakenly moved to private, despite being needed by a friend class.
Finalized usml_release_notes html file.