Corrected too shallow logic and updated make_tar
Too shallow logic was supposed to turn the ray around when the ray was in too shallow of water.
We need to remember this turn around logic twice, where previously it was only being considered
once. We needed to check for too shallow once when we compute dot_full and secondly after we
determine the exact location of the bottom interaction.

Added new exclude files for the make_tar to bring the tar size within reasonable size.